Modern_Cave_Paintings
The idea behind this project was to recreate one of the earliest forms of human expression and more specifically draw attention to how these ancient forms of art grew over time to tell a larger narrative with many authors.

The interaction is simple; you enter a simple expression to describe something interesting you've done today and the program will generate an image in the style of a modern cave painting and add it to the shared canvas. All of the collected images are stored and remain persistent for all users to see on the canvas as they add their own.